Added some missing commas in the witness graph definition and updated the test result.
In puzzle.py, correct the exit finding by adding the key, not (every letter of the) exit node.
This bug was not apparent from the test data due to the naming convention of the exit nodes
(e.g. "aExit") and the fact that no nodes were named 'E', 'x', 'i', or 't'.